+++ client/src/shared/hooks/useCreateCourse.ts
@@ -1,44 +1,31 @@
 /* eslint-disable @typescript-eslint/restrict-template-expressions */
 /* eslint-disable @typescript-eslint/no-floating-promises */
-import { useEffect, useState } from 'react';
-import type { SelectOptionData } from '@/src/shared/components/Select';
 import { useRouter } from 'next/router';
+import { useEffect, useState } from 'react';
+import { type SubmitHandler, useForm } from 'react-hook-form';
+import {
+  type CourseFormInput,
+  type MultiSelectOptionData,
+  alertError,
+  alertSuccess,
+  isRequestOk
+} from '../utils';
 import API from '@/src/apis';
-import { alertError, alertSuccess, isRequestOk } from '../utils';
 
 export const useCreateCourse = (): any => {
   const router = useRouter();
   const [category, setCategory] = useState([]);
 
-  // created_by is set to admin for now, will update soon after configuration of aut
-  const [postData, setPostData] = useState({
-    title: '',
-    description: '',
-    created_by: 1,
-    course_category: ''
-  });
-
-  const handleInput = (e: any): void => {
-    e.persist();
-    setPostData({ ...postData, [e.target.name]: e.target.value });
-  };
-
-  const handleCancel = (): void => {
-    router.back();
-  };
-
-  const handleSubmit = async (
-    e: React.FormEvent<HTMLFormElement> | React.MouseEvent<HTMLButtonElement>
-  ): Promise<void> => {
-    e.preventDefault();
-    try {
-      const result = await API.post('/course/', postData);
-      alertSuccess('Course Successfully Added');
-      router.push(`/trainer/course/detail/${result.data.id}`);
-    } catch (error) {
-      alertError('Something Went Wrong');
+  const {
+    register,
+    control,
+    handleSubmit,
+    formState: { errors }
+  } = useForm<CourseFormInput>({
+    defaultValues: {
+      category: []
     }
-  };
+  });
 
   const paths = [
     {
@@ -50,6 +37,38 @@ export const useCreateCourse = (): any => {
       url: router.asPath
     }
   ];
+
+  const handleCancel = (): void => {
+    router.back();
+  };
+
+  const categoriesOption: MultiSelectOptionData[] = Object.entries(
+    category
+  ).map(([key, { id, name }]) => ({
+    value: id,
+    label: name
+  }));
+
+  const onSubmit: SubmitHandler<CourseFormInput> = async (
+    data: CourseFormInput
+  ): Promise<void> => {
+    // Author and Company is still static due to verification of its point of origin (not discussed yet)
+    const postData = {
+      name: data.name,
+      category: data.category.map((option) => option.value).join(','),
+      description: data.description,
+      company: 1,
+      author: 1
+    };
+
+    try {
+      const result = await API.post('/course/', postData);
+      alertSuccess('Course Successfully Added');
+      router.push(`/trainer/course/detail/${result.data.id}`);
+    } catch (error) {
+      alertError('Something Went Wrong');
+    }
+  };
 
   useEffect(() => {
     const fetchData = async (): Promise<void> => {
@@ -65,22 +84,17 @@ export const useCreateCourse = (): any => {
       }
     };
 
-    fetchData();
+    void fetchData();
   }, []);
 
-  const categoriesOption: SelectOptionData[] = Object.entries(category).map(
-    ([key, { id, name }]) => ({
-      id,
-      text: name
-    })
-  );
-
   return {
-    handleInput,
-    handleCancel,
+    register,
+    control,
     handleSubmit,
-    postData,
+    errors,
     paths,
+    handleCancel,
+    onSubmit,
     categoriesOption
   };
 };
